create input components
clean up ai-gen and make jokes page a component

fix build errors
