+++ app/components/NavBar.tsx
@@ -12,7 +12,6 @@ function NavBar() {
             <Nav.Link href="/gallery">Gallery</Nav.Link>
             <Nav.Link href="/profile">Profile</Nav.Link>
             <Nav.Link href="/ai-gen">AI-Gen</Nav.Link>
-            <Nav.Link href="/jokes">Jokes</Nav.Link>
           </Nav>
         </Navbar.Collapse>
     </Navbar>
